requirements.txt
discord.py
gino
